refactor : responseDTO부분 정리 #86
